+++ src/persistence/serviceProvider.entity.ts
@@ -1,6 +1,6 @@
-import { TimestampedEntity } from './timestamped.entity.js';
+import { TimestampedEntity } from './timestamped.entity';
 import { Entity, ManyToOne, PrimaryKey } from '@mikro-orm/core';
-import { SchoolStructureNodeEntity } from './schoolStructureNode.entity.js';
+import { SchoolStructureNodeEntity } from './schoolStructureNode.entity';
 
 @Entity({ tableName: 'service_provider' })
 export class ServiceProviderEntity extends TimestampedEntity<ServiceProviderEntity, 'id'> {
