+++ src/persistence/serviceProviderAccess.entity.ts
@@ -1,5 +1,5 @@
 import { Entity } from '@mikro-orm/core';
-import { RolePermissionEntity } from './rolePermission.entity.js';
+import { RolePermissionEntity } from './rolePermission.entity';
 
 @Entity({ discriminatorValue: 'serviceProviderAccess' })
 export class ServiceProviderAccessEntity extends RolePermissionEntity {}
