+++ src/persistence/serviceProviderLicence.entity.ts
@@ -1,7 +1,7 @@
-import { TimestampedEntity } from './timestamped.entity.js';
+import { TimestampedEntity } from './timestamped.entity';
 import { Entity, ManyToOne, PrimaryKey } from '@mikro-orm/core';
-import { ServiceProviderEntity } from './serviceProvider.entity.js';
-import { SchoolStructureNodeEntity } from './schoolStructureNode.entity.js';
+import { ServiceProviderEntity } from './serviceProvider.entity';
+import { SchoolStructureNodeEntity } from './schoolStructureNode.entity';
 
 @Entity({ tableName: 'service_provider_licence' })
 export class ServiceProviderLicenceEntity extends TimestampedEntity<ServiceProviderLicenceEntity, 'id'> {
