+++ src/persistence/systemAccess.entity.ts
@@ -1,5 +1,5 @@
 import { Entity } from '@mikro-orm/core';
-import { RolePermissionEntity } from './rolePermission.entity.js';
+import { RolePermissionEntity } from './rolePermission.entity';
 
 @Entity({ discriminatorValue: 'systemAccess' })
 export class SystemAccessEntity extends RolePermissionEntity {}
