+++ src/persistence/userRoleAssignment.entity.ts
@@ -1,7 +1,7 @@
-import { TimestampedEntity } from './timestamped.entity.js';
+import { TimestampedEntity } from './timestamped.entity';
 import { Entity, ManyToOne, PrimaryKey } from '@mikro-orm/core';
-import { PersonEntity } from './person.entity.js';
-import { RoleEntity } from './role.entity.js';
+import { PersonEntity } from '../modules/person/persistence/person.entity';
+import { RoleEntity } from './role.entity';
 
 @Entity({ tableName: 'user_role_assignment' })
 export class UserRoleAssignmentEntity extends TimestampedEntity<UserRoleAssignmentEntity, 'id'> {
